+++ src/DesktopGrass.Win2D/TrayIcon.cs
@@ -88,7 +88,8 @@ internal sealed class TrayIcon : IDisposable
         var critterNoneItem  = new ToolStripMenuItem("None")  { Tag = CritterKind.None,  CheckOnClick = false };
         var critterSheepItem = new ToolStripMenuItem("Sheep") { Tag = CritterKind.Sheep, CheckOnClick = false };
         var critterCatItem   = new ToolStripMenuItem("Cat")   { Tag = CritterKind.Cat,   CheckOnClick = false };
-        var critterItems = new[] { critterNoneItem, critterSheepItem, critterCatItem };
+        var critterAllItem   = new ToolStripMenuItem("All")   { Tag = CritterKind.Bunny, CheckOnClick = false };
+        var critterItems = new[] { critterNoneItem, critterSheepItem, critterCatItem, critterAllItem };
 
         var petCountMenu = new ToolStripMenuItem("Pet count");
         var petCountRandomItem = new ToolStripMenuItem("Random") { Tag = 0, CheckOnClick = false };
@@ -116,6 +117,7 @@ internal sealed class TrayIcon : IDisposable
         critterNoneItem.Click  += (_, _) => SelectCritter(CritterKind.None);
         critterSheepItem.Click += (_, _) => SelectCritter(CritterKind.Sheep);
         critterCatItem.Click   += (_, _) => SelectCritter(CritterKind.Cat);
+        critterAllItem.Click   += (_, _) => SelectCritter(CritterKind.Bunny);
         foreach (var it in petCountItems)
         {
             int n = (int)it.Tag!;